fix: Consolidate Supabase client initialization
Resolves the 'Multiple GoTrueClient instances' warning by ensuring the Supabase client is initialized only once.

The redundant client at `src/lib/supabaseClient.ts` has been removed, and all imports have been updated to point to the canonical client at `src/integrations/supabase/client.ts`.

This change is intended to fix a black screen issue on the login page that was likely caused by conflicting client initializations.
