+++ src/pages/Integrations.tsx
@@ -4,7 +4,7 @@ import { Badge } from "@/components/ui/badge";
 import SocialAnalytics from "@/components/SocialAnalytics";
 import { Button } from "@/components/ui/button";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import { supabase } from "@/lib/supabaseClient";
+import { supabase } from "@/integrations/supabase/client";
 
 const Integrations = () => {
   const queryClient = useQueryClient();
